@@ -3,7 +3,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.neenbedankt.gradle.plugins:android-apt:1.4'
+        classpath 'com.neenbedankt.gradle.plugins:android-apt:1.8'
     }
 }
 
@@ -12,33 +12,32 @@ repositories {
 }
 
 apply plugin: 'com.android.application'
-apply plugin: 'com.neenbedankt.android-apt'
+//apply plugin: 'com.neenbedankt.android-apt'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion '25.0.2'
+    compileSdkVersion 30
+    buildToolsVersion '29.0.3'
 
     defaultConfig {
         applicationId 'com.danikula.videocache.sample'
         minSdkVersion 15
-        targetSdkVersion 23
+        targetSdkVersion 30
         versionCode 1
         versionName '1.0'
     }
 }
 
-apt {
-    arguments {
-        androidManifestFile variant.outputs[0].processResources.manifestFile
-        resourcePackageName android.defaultConfig.applicationId
-    }
-}
+//apt {
+//    arguments {
+//        androidManifestFile variant.outputs[0].processResources.manifestFile
+//        resourcePackageName android.defaultConfig.applicationId
+//    }
+//}
 
 dependencies {
-//    compile project(':library')
-    compile 'com.android.support:support-v4:23.1.0'
-    compile 'org.androidannotations:androidannotations-api:3.3.2'
-    compile 'com.danikula:videocache:2.7.1'
-    compile 'com.viewpagerindicator:library:2.4.2-SNAPSHOT@aar'
-    apt 'org.androidannotations:androidannotations:3.3.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'org.androidannotations:androidannotations-api:4.7.0'
+    implementation 'com.viewpagerindicator:library:2.4.2-SNAPSHOT@aar'
+    implementation project(path: ':library')
+    annotationProcessor 'org.androidannotations:androidannotations:4.7.0'
 }
