@@ -1,8 +1,9 @@
 package com.danikula.videocache.sample;
 
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v4.app.FragmentActivity;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.FragmentActivity;
+
 import android.util.Log;
 import android.widget.ArrayAdapter;
 import android.widget.ListAdapter;
