@@ -1,7 +1,7 @@
 package com.danikula.videocache.sample;
 
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
 
 import org.androidannotations.annotations.EActivity;
 
