@@ -2,10 +2,10 @@ package com.danikula.videocache.sample;
 
 import android.os.Handler;
 import android.os.Message;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.util.Log;
 import android.widget.ImageView;
-import android.widget.ProgressBar;
+import android.widget.SeekBar;
 import android.widget.VideoView;
 
 import com.danikula.videocache.CacheListener;
@@ -28,7 +28,7 @@ public class VideoFragment extends Fragment implements CacheListener {
 
     @ViewById ImageView cacheStatusImageView;
     @ViewById VideoView videoView;
-    @ViewById ProgressBar progressBar;
+    @ViewById SeekBar progressBar;
 
     private final VideoProgressUpdater updater = new VideoProgressUpdater();
 
@@ -78,7 +78,9 @@ public class VideoFragment extends Fragment implements CacheListener {
     public void onDestroy() {
         super.onDestroy();
 
-        videoView.stopPlayback();
+        if (videoView!=null) {
+            videoView.stopPlayback();
+        }
         App.getProxy(getActivity()).unregisterCacheListener(this);
     }
 
